fix: reorganizando a estrutura dos componentes
