+++ src/App.jsx
@@ -2,29 +2,39 @@ import React from "react";
 import {
   RouterProvider,
   createBrowserRouter,
+  createRoutesFromElements,
+  Route,
+  Outlet
 } from "react-router-dom";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
-import HomePage from "./components/HomePage";
-import Agenda from "./components/Agenda";
-import Contact from "./components/Contact";
+import HomePage from "./components/pages/HomePage";
+import Agenda from "./components/pages/Agenda";
+import Contact from "./components/pages/Contact";
+import ScrollToTop from "./ScrollToTop"; // Import the custom component
 
 const queryClient = new QueryClient();
 
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <HomePage />,
-  },
-  {
-    path: "/agenda",
-    element: <Agenda />,
-  },
-  {
-    path: "/contact",
-    element: <Contact />,
-  },
-]);
+// Layout component that includes ScrollToTop and renders children
+function Layout() {
+  return (
+    <>
+      <ScrollToTop />
+      <Outlet />
+    </>
+  );
+}
+
+// Create routes with the Layout component as the root
+const router = createBrowserRouter(
+  createRoutesFromElements(
+    <Route element={<Layout />}>
+      <Route path="/" element={<HomePage />} />
+      <Route path="/agenda" element={<Agenda />} />
+      <Route path="/contact" element={<Contact />} />
+    </Route>
+  )
+);
 
 function App() {
   return (
